Final Corrections

+++ KmandiliWebService/Controllers/ApplicationControllers/PastryShopsController.cs
@@ -1,17 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Data;
+﻿using System.Collections.Generic;
 using System.Data.Entity;
 using System.Data.Entity.Infrastructure;
-using System.Data.Entity.Validation;
 using System.IO;
 using System.Linq;
 using System.Net;
-using System.Net.Http;
 using System.Web.Hosting;
 using System.Web.Http;
 using System.Web.Http.Description;
-using System.Web.Http.Results;
 using KmandiliWebService.DatabaseAccessLayer;
 
 namespace KmandiliWebService.Controllers.ApplicationControllers
@@ -19,19 +14,19 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
     [Authorize]
     public class PastryShopsController : ApiController
     {
-        private KmandiliDBEntities db = new KmandiliDBEntities();
+        private readonly KmandiliDBEntities _db = new KmandiliDBEntities();
 
         // GET: api/PastryShops
         public IQueryable<PastryShop> GetPastryShops()
         {
-            return db.PastryShops;
+            return _db.PastryShops;
         }
 
         // GET: api/PastryShops/5
         [ResponseType(typeof(PastryShop))]
         public IHttpActionResult GetPastryShop(int id)
         {
-            PastryShop pastryShop = db.PastryShops.Find(id);
+            PastryShop pastryShop = _db.PastryShops.Find(id);
             if (pastryShop == null)
             {
                 return NotFound();
@@ -44,7 +39,7 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
         [ResponseType(typeof(PastryShop))]
         public IHttpActionResult GetPastryShop(string email, string password)
         {
-            PastryShop pastryShop = db.PastryShops.FirstOrDefault(p => p.Email == email && p.Password == password);
+            PastryShop pastryShop = _db.PastryShops.FirstOrDefault(p => p.Email == email && p.Password == password);
             if (pastryShop == null)
             {
                 return NotFound();
@@ -58,7 +53,7 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
         [ResponseType(typeof(PastryShop))]
         public IHttpActionResult GetPastryShop(string email)
         {
-            PastryShop pastryShop = db.PastryShops.FirstOrDefault(p => p.Email == email);
+            PastryShop pastryShop = _db.PastryShops.FirstOrDefault(p => p.Email == email);
             if (pastryShop == null)
             {
                 return NotFound();
@@ -82,21 +77,22 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
                 return BadRequest();
             }
 
-            var existingPastry = db.PastryShops.FirstOrDefault(p => p.ID == id);
-            var toDeleteCategories = existingPastry.Categories.Except(pastryShop.Categories,c => c.ID).ToList<Category>();
-            var toAddCategories = pastryShop.Categories.Except(existingPastry.Categories,c=> c.ID).ToList<Category>();
+            var existingPastry = _db.PastryShops.FirstOrDefault(p => p.ID == id);
+            if (existingPastry == null) return NotFound();
+            var toDeleteCategories = existingPastry.Categories.Except(pastryShop.Categories,c => c.ID).ToList();
+            var toAddCategories = pastryShop.Categories.Except(existingPastry.Categories,c=> c.ID).ToList();
 
             toDeleteCategories.ForEach(c => existingPastry.Categories.Remove(c));
             foreach (var category in toAddCategories)
             {
-                if (db.Entry(category).State == EntityState.Detached)
-                    db.Categories.Attach(category);
+                if (_db.Entry(category).State == EntityState.Detached)
+                    _db.Categories.Attach(category);
                 existingPastry.Categories.Add(category);
             }
 
             try
             {
-                db.SaveChanges();
+                _db.SaveChanges();
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -104,10 +100,7 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
                 {
                     return NotFound();
                 }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
             return StatusCode(HttpStatusCode.NoContent);
         }
@@ -126,11 +119,11 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
                 return BadRequest();
             }
             
-            db.Entry(pastryShop).State = EntityState.Modified;
+            _db.Entry(pastryShop).State = EntityState.Modified;
 
             try
             {
-                db.SaveChanges();
+                _db.SaveChanges();
             }
             catch (DbUpdateConcurrencyException)
             {
@@ -138,10 +131,7 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
                 {
                     return NotFound();
                 }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
 
             return StatusCode(HttpStatusCode.NoContent);
@@ -159,13 +149,10 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
 
             foreach (Category pastryShopCategory in pastryShop.Categories)
             {
-                db.Entry(pastryShopCategory).State = EntityState.Unchanged;
+                _db.Entry(pastryShopCategory).State = EntityState.Unchanged;
             }
-            db.PastryShops.Add(pastryShop);
-            db.SaveChanges();
-
-            //db.PastryShops.Add(pastryShop);
-            //db.SaveChanges();
+            _db.PastryShops.Add(pastryShop);
+            _db.SaveChanges();
 
             return CreatedAtRoute("DefaultApi", new { id = pastryShop.ID }, pastryShop);
         }
@@ -174,23 +161,26 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
         [ResponseType(typeof(PastryShop))]
         public IHttpActionResult DeletePastryShop(int id)
         {
-            PastryShop pastryShop = db.PastryShops.Find(id);
+            PastryShop pastryShop = _db.PastryShops.Find(id);
             if (pastryShop == null)
             {
                 return NotFound();
             }
             string profileFileName = pastryShop.ProfilePic.Substring(44, (pastryShop.ProfilePic.Length - 44));
             string profileImageFilePath = HostingEnvironment.MapPath("~/Uploads");
+            if (profileImageFilePath == null) return NotFound();
             profileImageFilePath = Path.Combine(profileImageFilePath, profileFileName);
 
             string coverFileName = pastryShop.CoverPic.Substring(44, (pastryShop.CoverPic.Length - 44));
             string coverImageFilePath = HostingEnvironment.MapPath("~/Uploads");
+            if (coverImageFilePath == null) return NotFound();
             coverImageFilePath = Path.Combine(coverImageFilePath, coverFileName);
             
             foreach (var product in pastryShop.Products)
             {
                 string productFileName = product.Pic.Substring(44, (product.Pic.Length - 44));
                 string productImageFilePath = HostingEnvironment.MapPath("~/Uploads");
+                if (productImageFilePath == null) return NotFound();
                 productImageFilePath = Path.Combine(productImageFilePath, productFileName);
                 File.Delete(productImageFilePath);
             }
@@ -203,13 +193,15 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
 
             var phoneNumbers = new List<PhoneNumber>(pastryShop.PhoneNumbers);
             var orders = new List<Order>(pastryShop.Orders);
-            db.PastryShops.Remove(pastryShop);
-            phoneNumbers.ForEach(p => db.PhoneNumbers.Remove(p));
-            poinOfSalesPhoneNumbers.ForEach(p => db.PhoneNumbers.Remove(p));
-            orders.ForEach(o => db.Orders.Remove(o));
-            db.Addresses.Remove(db.Addresses.Find(pastryShop.Address_FK));
+            _db.PastryShops.Remove(pastryShop);
+            phoneNumbers.ForEach(p => _db.PhoneNumbers.Remove(p));
+            poinOfSalesPhoneNumbers.ForEach(p => _db.PhoneNumbers.Remove(p));
+            orders.ForEach(o => _db.Orders.Remove(o));
+            var a = _db.Addresses.Find(pastryShop.Address_FK);
+            if (a == null) return NotFound();
+            _db.Addresses.Remove(a);
             
-            db.SaveChanges();
+            _db.SaveChanges();
             File.Delete(profileImageFilePath);
             File.Delete(coverImageFilePath);
             return Ok(pastryShop);
@@ -219,14 +211,14 @@ namespace KmandiliWebService.Controllers.ApplicationControllers
         {
             if (disposing)
             {
-                db.Dispose();
+                _db.Dispose();
             }
             base.Dispose(disposing);
         }
 
         private bool PastryShopExists(int id)
         {
-            return db.PastryShops.Count(e => e.ID == id) > 0;
+            return _db.PastryShops.Count(e => e.ID == id) > 0;
         }
         
     }
